Projeto1_AutomacaoAnaliseDados/requirements.txt
# =============================================
# DEPENDÊNCIAS DO PROJETO
# Projeto 1: Automação e Análise de Dados com Python
# =============================================

# --- Automação (RPA) ---
pyautogui==0.9.54       # Controle de mouse e teclado

# --- Manipulação de Dados ---
pandas==3.0.2           # Leitura e análise de dados
openpyxl==3.1.5         # Leitura de arquivos .xlsx
numpy==2.4.4            # Operações numéricas

# --- Visualização ---
plotly==6.7.0           # Gráficos interativos

# --- Relatório (Quarto) ---
nbformat==5.10.4        # Suporte a notebooks Jupyter (usado pelo Quarto)
nbclient==0.10.4        # Execução de notebooks (usado pelo Quarto)
pyyaml==6.0.3           # Parser YAML (usado pelo Quarto)

# --- Configuração e Ambiente ---
python-dotenv==1.2.2    # Carregamento de variáveis de ambiente (.env)

# --- Desenvolvimento e Notebook ---
jupyter==1.1.1          # Ambiente Jupyter Notebook/Lab
ipykernel==7.2.0        # Kernel Python para Jupyter
